Reorganise helper functions for conversation notifications (#1588)
* Break up `addToConversation` to its constituents
* Move Galley helper functions to a `Util` module
* Use `runFederated` to avoid code duplication
